[이현호] 문서 수정 및 삭제 조건 분기 처리
- 문서 수정
	- 결재 대기 상태라도 결재자가 존재하면 수정 불가 처리
	- 승인된 문서 수정 불가 처리
- 문서 삭제
	- '승인/반려' 상태인 경우 삭제 불가 처리
	- 결재자가 존재할 시(FK 제약 위반 발생 시) 삭제 불가 처리
- 사인 상태 변경 시 REJECTED일 경우에만 사인 정보 삭제하도록 서비스 로직 수정

+++ signapp/src/main/java/com/example/signapp/service/DocService.java
@@ -47,17 +47,21 @@ public class DocService {
 	public int deleteDocument(int docNo) {
 		return docMapper.deleteDocument(docNo);
 	}
+	
 	public int updateSignStatus(int docNo, String signStatus) {
-		// 반려면 그 문서에 있는 사인들도 모두 지워줘야함...
 		int row =  docMapper.updateSignStatus(docNo, signStatus);
-		if (row > 0) {
+		
+		// 반려일 때만 사인 삭제
+		if (row > 0 && "REJECTED".equals(signStatus)) {
 			signMapper.deleteSign(docNo);
 		}
 		return row;
 	}
+	
 	public List<Document> selectDocumentList(Page page) {
 		return docMapper.selectDocumentList(page);
 	}
+	
 	public int totalCount(Page page) {
 		return docMapper.totalCount(page);
 	}
